requirements.txt
requests
beautifulsoup4
sentence-transformers
faiss-cpu
numpy
pandas
flask
flask-cors
pyngrok
openai
dotenv
